turiniq-be/requirements.txt
google-adk
google-generativeai
python-dotenv
fastapi
uvicorn
pymongo
requests
beautifulsoup4
pypdf2
pydantic==2.11.0
websockets